added type and move text label into language file

+++ packages/client-core/src/user/components/UserMenu/menus/AvatarMenu.tsx
@@ -14,6 +14,7 @@ import { LazyImage } from '../../../../common/components/LazyImage'
 import { AuthService, useAuthState } from '../../../services/AuthService'
 import styles from '../UserMenu.module.scss'
 import { Views } from '../util'
+import { UserAvatar } from '@xrengine/common/src/interfaces/UserAvatar'
 
 interface Props {
   changeActiveMenu: Function
@@ -34,7 +35,7 @@ const AvatarMenu = (props: Props) => {
   const [imgPerPage, setImgPerPage] = useState(getAvatarPerPage())
   const [selectedAvatarId, setSelectedAvatarId] = useState('')
   const [isAvatarLoaded, setAvatarLoaded] = useState(false)
-  const [avatarTobeDeleted, setAvatarTobeDeleted] = useState<any>(null!)
+  const [avatarTobeDeleted, setAvatarTobeDeleted] = useState<UserAvatar | null>()
 
   let [menuRadius, setMenuRadius] = useState(window.innerWidth > 360 ? 182 : 150)
 
@@ -101,11 +102,11 @@ const AvatarMenu = (props: Props) => {
     setPage(page - 1)
   }
 
-  const selectAvatar = (avatarResources: any) => {
+  const selectAvatar = (avatarResources: UserAvatar) => {
     const avatar = avatarResources.avatar
-    setSelectedAvatarId(avatar.name)
-    if (avatarId !== avatar.name) {
-      setAvatar(avatar.name, avatar.url, avatarResources['user-thumbnail'].url)
+    setSelectedAvatarId(avatar?.name || '')
+    if (avatarId !== avatar?.name) {
+      setAvatar(avatar?.name || '', avatar?.url || '', avatarResources?.userThumbnail?.url || '')
     }
   }
 
@@ -126,7 +127,7 @@ const AvatarMenu = (props: Props) => {
 
   const removeAvatar = (e, confirmation) => {
     e.stopPropagation()
-    if (confirmation) {
+    if (confirmation && avatarTobeDeleted?.avatar?.key) {
       AuthService.removeAvatar(avatarTobeDeleted.avatar.key)
     }
 
@@ -191,9 +192,9 @@ const AvatarMenu = (props: Props) => {
           }}
         >
           <CardContent onClick={() => selectAvatar(characterAvatar)}>
-            <LazyImage key={avatar.id} src={characterAvatar['user-thumbnail'].url} alt={avatar.name} />
+            <LazyImage key={avatar.id} src={characterAvatar?.userThumbnail?.url || ''} alt={avatar.name} />
             {avatar.userId ? (
-              avatarTobeDeleted && avatarTobeDeleted.avatar.url === avatar.url ? (
+              avatarTobeDeleted && avatarTobeDeleted?.avatar?.url === avatar.url ? (
                 <div className={styles.confirmationBlock}>
                   <p>{t('user:usermenu.avatar.confirmation')}</p>
                   <button
